# Optimize `ReadCacheLookup`, 6.7 speed-up