--- a/Cargo.toml
+++ b/Cargo.toml
@@ -77,6 +77,7 @@ k256 = { version = "0.13.4", features = ["ecdsa"] }
 p256 = { version = "0.13.2", features = ["ecdsa"] }
 # for keccak256
 sha3 = "0.10.8"
+rand = { workspace = true }
 
 [dev-dependencies]
 rstest = { workspace = true }