--- a/Cargo.toml
+++ b/Cargo.toml
@@ -61,6 +61,7 @@ serde = "1.0.214"
 serde_json = "1.0.133"
 clap = "4.5.20"
 rand_chacha = "0.3.1"
+bitvec = "1.0.1"
 
 [dependencies]
 lazy_static = { workspace = true }
@@ -70,6 +71,7 @@ num-integer = { workspace = true }
 chia-bls = { workspace = true }
 chia-sha2 = { workspace = true }
 hex-literal = { workspace = true }
+bitvec = { workspace = true }
 # for secp sigs
 k256 = { version = "0.13.4", features = ["ecdsa"] }
 p256 = { version = "0.13.2", features = ["ecdsa"] }