@arvidn 2024-12-29
a material portion of the speed-up in this patch comes from not using a salt in the hasher. Which is a bit surprising, but I suppose we initialize new `HashSet` and `HashMaps` quite often. adding back the salt makes the benchmarks take 3x the time